requirements.txt
google_api_python_client==2.105.0
google_auth_oauthlib==1.1.0
matplotlib==3.7.2
pandas==2.0.3
protobuf==4.24.4
python_crontab==3.0.0
